Add A Fine Mist of Blood
I couldn't find a feed, only soundcloud so I was making a feed.
Then I found panoply with the first link, but still no feed. Then
through that page an jsonp file (changed the format to json
straight away) with their podcasts, and itunes link (that happened
to me null) and finally a rss feed. But it was feed burner so I
made it anyway.

http://podcasts.dev.slate.com/cms/api/v1/show/?format=jsonp&featured=true&order_by=priority&callback=importShows&limit=1000;
http://podcasts.dev.slate.com/cms/api/v1/show/?format=json&featured=true&order_by=priority&limit=1000;
http://feeds.feedburner.com/BoschAFineMistOfBlood